pyproject.toml
[project]
name = "sm-13"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "kubernetes>=34.1.0",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.0",
    "mkdocs-mermaid2-plugin>=1.2.3",
    "pymdown-extensions>=10.19.1",
]
